project_choa/pyproject.toml
[project]
name = "project-choa"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "langchain-openai (>=0.3.27,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "aiogram (>=3.21.0,<4.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]

[tool.poetry]
packages = [{include = "project_choa", from = "src"}]


[tool.poetry.group.dev.dependencies]
notebook = "^7.4.4"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
